Fix DCE force live locations for multi-pred + multi-succ
Summary:
D13105801 added a fixed this issues for blocks multi pred blocks that were a
direct successor of a multi successor block.  This fixes that to work in cases
where there are blocks along the path between the muti pred block and the multi
succ block.

This adds a copy of `hphp/test/slow/hhbbc/global-dce.hhas` with some Nop blocks
along critical edges.  It causes DCE to make some in correct inferences about
stack usages.  This diff fixes it.

Reviewed By: ricklavoie

Differential Revision: D21066125

fbshipit-source-id: 83c2cd4f5891ab80afcce58b8568041c0d8ed795