Kill frame locals in memory-effects even in elided frames
Summary:
Right now the memory-effects for InlineReturnNoFrame does not
kill the locals for the frame. This is because with the way we do
frame elision right now, we shouldn't have any accesses to the frame's
locals if we elided it. Local accesses should be rewritten to be stack
accesses.

I am changing this, since we will not be able to masquerade local
accesses as stack accesses (they will have a different format). So,
instead of changing to stack accesses, I will instead rewrite the
local access as using a parent frame pointer, with an additional
static offset.

Since we will now be accessing the elided frame's locals, we need them
to be killed properly when we return from the frame. Otherwise
store-elim will sink the stores to after the return, and we'll write
to something else.

So, InlineReturnNoFrame needs to have the same memory-effects as
InlineReturn. The only difficulty is we don't have a frame pointer to
calculate the needed offsets from. Changing InlineReturnNoFrame to
take the StkPtr and contain a SPRelOffset to where the frame would
start if it hadn't been elided. This information is easily available
from the DefInlineFP when we convert the InlineReturn into an
InlineReturnNoFrame. Using the SPRelOffset is convenient because it
does not need to be updated if we elide more frames (unlike the frame
pointer relative offsets). Using this offset to calculate the
FPRelOffset of where the frame would be, and use that in
memory-effects.

I also think that the implementation of InlineReturnNoFrame was wrong,
as it was calculating its offset from the outermost frame, but was
using rvmfp() as the base (which is not necessarily the outermost
frame). This only matters if we have HHIR assertions enabled, and I
think is benign anyways. Regardless, this is now fixed.

Reviewed By: jano

Differential Revision: D21085740

fbshipit-source-id: ce3d2c0054b1c04a7d51a7baf61639537090b2a1