Clean up may-reenter logic in memory-effects
Summary:
Historically in memory-effects, we needed to add may_reenter in two
cases. The first case is if the instruction might raise an error
(meaning it has a catch block), since any re-entry could possibly
throw. The second case is if the instruction would perform a
dec-ref. Dec-refs were the only situation where we could re-enter
(because of destructors), but not need an explicit catch block
(because destructors ate exceptions).

In memory-effects, the first case has long been dealt with
automatically. We automatically add in may_reenter if the instruction
can raise an error. The second case no longer applies because dec-refs
can no longer re-enter. So, there's no longer any need for an
instruction to manually call may_reenter.

Remove the manual calls to may_reenter. They either don't apply
because dec-refs don't re-enter anymore, or they were already being
applied because the instructions raise. We no longer need the special
cases check in may_reenter. Also remove some now un-true or outdated
comments.

Since we no longer have a set of instructions which can re-enter
without raising, we no longer need the special logic in frame dead
code elimination to fixup the spOff of those instructions. We
shouldn't be eliminating frames of potentially re-entering
instructions, so assert that.

IterInit and IterInitK had special logic to not add may_reenter if the
inputs weren't an Obj. This actually didn't do anything because
IterInit and IterInitK are marked as raising an error, thus got the
treatment automatically. Add a new predicate to IRInstruction which
checks if the instruction might raise, taking into account input
types. Use that in memory effects instead to avoid adding
may_reenter. We can't use it everywhere because the result may change
as types are flowed.

Reviewed By: ottoni, jano

Differential Revision: D21085227

fbshipit-source-id: d9104f93f72e0c469f48c1e96f67261ebc7710c9