Remove Typename from the AAST
Summary:
The `Typename` node is a node created during the naming.ml transform phase specifically for `Foo::class` when `Foo` is a typedef.

As it relies on global information (as `Foo` may not be a typedef defined locally, it seems like an extreme transformation for naming to make. The benefits for typing are not very large, so instead of moving the transformation later down the pipeline or earlier up the pipeline - this diff removes in entirely.

Reviewed By: Wilfred

Differential Revision: D21003641

fbshipit-source-id: e304f68f55cf440491ac20481ed9737e1d1d05cf