Change ArrayIter log flag to bool
Summary:
There are only a few 10k of loops in the production TC. If we are going to log information about these loops in some small test (e.g. as a way to gather data for further optimizations), there's no reason not to log all of them.

As a follow-up, I think we should add a logging flag to all TargetProfiles.

Reviewed By: ricklavoie

Differential Revision: D21103332

fbshipit-source-id: c19e2e683c4e2653ec110ab70f1401e11a114b5e