When removing frames, rewrite local accesses using parent FP plus static offset
Summary:
Right now, when attempting to remove inlined frames, we attempt to
change accesses to locals to instead be stack accesses (this drops the
requirement of having a FramePtr). This only works because locals and
stack elements have the same layout, and therefore we can pun them as
each other.

To remove this restriction, rewrite the instructions a different
way. Instead of converting it to a stack access, change it to instead
use a parent FramePtr, plus a static offset. All these instructions
only use the FramePtr to calculate the address of the local. So, it
can use a parent FramePtr, plus an offset from the parent FramePtr to
arrive at the same address. The advantage is that we are now free to
modify the layout of locals separate from stack elements.

This approach has one downside: We cannot perform this transformation
if the parent FramePtr is for a resumed frame, as there isn't a static
offset from the parent to the child in that case. We are forced to
keep the DefInlineFP for the first child. This does not seem to be
significant in practice.

Modify the logic in pdce-inlining and dce to do this new
transformation, and remove the convert to stack logic. Add a
FPInvOffset to all the relevant instructions, and use that offset
during code-gen. Assert in a variety of places that the offset is
always zero when the instruction is gened, as it should only be
non-zero after optimization (and frame-state doesn't know how to deal
with it). The memoization instructions get a similar treatment.

Reviewed By: paulbiss, mofarrell

Differential Revision: D21125511

fbshipit-source-id: 6cdd676646a37c59d988b6b4c91b1166a266f8bc