Prefer registers that don't require REX prefix in the graph-coloring allocator
Summary:
On X64, using the extended registers (x8-x15) requires a REX prefix,
and therefore should be avoided.  I noticed a lot of uses of the
extended registers in cases where some of the older, original x86
registers were unused.  I tracked it down to some logic in the
register allocator that was trying to randomly break ties.  This
diff drops that logic, thus breaking ties based on the register
number, which favors the original registers over the extended ones.

Reviewed By: ricklavoie, paulbiss

Differential Revision: D21146715

fbshipit-source-id: 388baec88fc8aeb548942e7fc2e4a407760f4b31