Make initialized property check see if type is a supertype of null
Summary:
The check for uninitialized properties is very incomplete. Currently, the criteria for allowing a property to skip initialization are:

```
    match sp.sp_type with
    | None -> false
    | Some ty ->
      (match get_node ty with
      | Tprim Tnull
      | Toption _
      | Tmixed ->
        false
```
In other words, missing type, `null` type, any `?` type, and mixed. Predictably, this fails even if the property's type is an alias to these listed types, as in the test case I added.

Instead, I've changed the check to see if `null <: propertyType`. This covers `mixed`, `null`, and optional types, but also covers any equivalent types. It also allows for properties that are explicitly `Tany` via FIXME and alias magic to not be initialized because `null <: Tany` (Hack Foundation request).

Differential Revision: D20854709

fbshipit-source-id: c3cf287b97dd4f6e7cc556b2633583c4dd8e2dcb