Check arguments of calls on nothing
Summary:
When a call happens on an expression for which hh inferred the type nothing, the arguments are not visited by the typechecker. This can be confusing because obvious errors such as `1 + "foo"` or a static call on a method that does not exist are not flagged.

This diff fixes the confusing behavior by checking arguments of such calls against Tany. While the check will always succeed, some errors in the argument expressions are now flagged.

Reviewed By: Wilfred

Differential Revision: D20891078

fbshipit-source-id: d48eddd930fdac72affe721b6acb69c8f441dafd