Change the type of format %s from mixed to ?arraykey
Summary:
In the future, implicit string casts on arrays and objects will raise exceptions.

`Str\format` already expects `%s` to be a `string`:
https://github.com/hhvm/hsl/blob/665f470b93ad3dcc375b357f3d90d70c474e0746/src/str/format.php#L18

Ideally this would be changed to be `string` but that raises a ton of errors, and `?arraykey` is and will always be safe to coerce to string, so we can do this as an incremental step towards `string`.

Reviewed By: DavidSnider

Differential Revision: D20443602

fbshipit-source-id: 340645dd34cb5e3f32320af44dd262c61c806ca5