Add associative array and array-backed set
Summary:
I'm interested in experimenting with arena allocation in the direct-decl parser. However, decls (like many other Hack types) contain several maps and sets (often IMap/ISet/SMap/SSet). The arena-allocator library we are using, bumpalo, does not provide BTreeMap or BTreeSet--it provides only Vec and String.

One way of dealing with this is by writing a bump-allocator-based balanced-tree implementation (only after starting to write this diff did I notice that we had already done so in arena_collections!). But because the maps and sets in decls tend to be very small in size, I think we may get better performance from associative arrays, where better cache-efficiency may make up for the cost of linear-time lookups.

Differential Revision: D20905537

fbshipit-source-id: f2df4ef97e92d2ff5b75db9e0a680fe08917cf4b