Add ability to produce by-reference oxidized types
Summary:
I'd like to experiment with arena-allocation in the direct decl parser. To do so, I need by-reference versions of all the types in decls, which includes everything in Ty, and (unfortunately) everything in the AST (because const decls include the initializer expression).

This may also be useful for generating arena-allocated types for the typechecker, but I have considered typechecker types to be out of scope for the moment while I get things working for ASTs and decls.

Some oxidized types do not contain data which would need to be borrowed, so do not require a lifetime (for instance, aast::ReifyKind, an enum of three zero-argument variants). Instead of trying to track whether or not types need a lifetime in hh_oxidize, we have it accept an `--extern-types-file` argument for a file containing the names of types which do not need a lifetime, and add import statements for them (importing them from the `oxidized` crate) in the generated output.

Reviewed By: CatherineGasnier

Differential Revision: D20905992

fbshipit-source-id: df16d876ace089cb309a01741fbbd3c78787c38d