Remove record field type-overriding behavior
Summary: This behavior was added to change the type of the doc comment fields in the AAST (so as to redirect their representation to a manually-written module that adds an Rc). We can remove this special-casing if we define a Doc_comment module on the OCaml side (but do not oxidize it, using the manual Rust definition of DocComment instead).

Reviewed By: shiqicao

Differential Revision: D20930861

fbshipit-source-id: d8a6df4b14d1e04963725b6b3314f95b2414de7a