dvarray enforcement optimizations
Summary: It's not safe to turn on these optimizations without implementing the rest of the flag, but I checked on small test cases (which happen to not have dvarray type hint violations) that the results look good. I'll add unit tests when we get done with the enforcement side of the flag.

Reviewed By: ricklavoie

Differential Revision: D21163055

fbshipit-source-id: 7686a25b38707a441778649bc2c9c932f86c8272