Compute `Provider_context.entry` contents lazily
Summary:
We'll eventually want to identify entries by SHA1 hash and use cached data without having to hit disk. This commit makes it so that we don't eagerly read contents from disk when constructing an entry. Instead, the user will explicitly call `Provider_context.compute_contents` to do so.

This is a little bit troublesome since we're shifting the possible exception site from the `make_entry` calls to where we actually ask for the contents. I tried to wrap the calls to `compute_contents` with an exception handler where it seemed to be appropriate. In most of the cases, we're expecting the contents to have been populated by previous work (particularly serving an IDE query).

Reviewed By: ljw1004

Differential Revision: D21030814

fbshipit-source-id: 5005a319214bd1767a08b9199f5dabe0523fcd3e