Use the correct API to construct paths when updating the naming table in the "full-init" case
Summary:
This bug is the reason why I turned the `load_naming_table_on_full_init` feature off for full init + remote type checking.

The issue is that `from_root` does something other than what the name implies: it seems like you can give it an absolute path, and it should construct a relative path assuming that the prefix is `Root`, but it actually assumes you're giving it a relative path/suffix and prepending the root prefix to it. As a result, this function is constructing invalid paths if you give it absolute paths.

Reviewed By: yinghuitan

Differential Revision: D21170468

fbshipit-source-id: 83c4fb370e0cc6be724ac8749e0498ad65f0deb0