Raise on varray implicit append / promotion
Summary:
These two behaviors are easy to enforce, because the flags are only used in the runtime (and not in HHBBC, where we'd have to audit the usage in more detail).

I made the new behavior as similar to vecs as possible. For setting integer keys that are not in bounds, we raise IndexOOB error, like a vec. For setting string keys, we raise invalid array key error, like a vec. For unsetting, the vec error message includes the word "vec", so we create a new one that uses "varray" instead.

Unset behaves differently for a vector-like PHP arrays and vec unset behavior. For a PHP array, unset doesn't change the "next integer index", so if you unset the last element of a varray and then append, you'll get an array with non-contiguous integer keys. For a vec, if you unset and then append, you'll get the same vec as before. Because we log notices on ALL varray unsets, we can use them to change the varray behavior to match the vec behavior; we do so when the specialization flag is turned on.

Reviewed By: ricklavoie

Differential Revision: D21166115

fbshipit-source-id: 9469b56332e0d456c3a9532fbf5dd0d0dbc50765