thread env through ide_rpc
Summary:
This trivial diff just thread "env", the global immutable environment, through to ide_rpc.

(Why? because I'll want ide_rpc to be able to refresh status. And status involves closing over enough functions and variables to be able to support the user clicking "Restart Hack IDE". And naturally we need to know env, in order to start that.)

Reviewed By: arxanas

Differential Revision: D21119720

fbshipit-source-id: e6c4a4105d5c33da0f751502818c7bc5938e2a41