clean up "verbose_to_file"
Summary:
This is minor cleanup. Why did we pass "verbose_to_stderr" as part of server startup, but only send "verbose_to_file" via rpc? I'm just making this more regular.

(My ultimate goal is to show busy status while clientIdeService is doing an RPC. But to send busy status during rpc we need to provide the callback for what happens if the user clicks RestartIDE. And if the user clicks RestartIDE then we have to restart it. And prior to this diff, restarting it involved sending an rpc to it to establish its verbosity. With this diff, I break the cycle.)

Reviewed By: arxanas

Differential Revision: D21114541

fbshipit-source-id: 7794c4ab68073be8a029c24e2c0e07d39c5d6217