show as busy during slow compute_tast
Summary:
Wen you first start serverless-IDE, and it finishes loading saved-state, then it presents itself as "Ready" -- but the first action you do ends up taking 5-10 seconds to populate the decl cache.

I worried that this is so confusing to users that I called it out as the only "known bug" in the HackIDE release announcement.

This diff changes it to show as busy during RPC calls. It turns on the busy indicator once it's taken longer than 200ms (to avoid flickering on and off during routine use), and turns it off as soon as it's done.

I did this with clientIdeService. That's because we don't have pre-emption inside clientIdeDaemon, hence no way for it to report 200ms into its compute_tast operation.

Reviewed By: arxanas

Differential Revision: D21119740

fbshipit-source-id: bf3005364cda74d9248dab70544a9c3032fa1b4c