new module clientIdeUtils
Summary:
This series of three diffs is to pay down some tech debt that I accumulated.

For a bug, we have three things we want to  do with it: (1) log it, (2) produce a user-facing error message in the status bar tooltip, (3) produce a developer-facing Lsp.Error jsonrpc response.

This diff just gathers together those three functions into a new module, clientIdeUtils. It will  be consumed by both clientIdeDaemon and clientIdeService because they both need to report the same bugs in the same way.

Reviewed By: arxanas

Differential Revision: D21148912

fbshipit-source-id: 8da32212ed86221aea8296106c3d8292a3e825b3