refactor clientLsp errors
Summary:
This diff pays back some tech debt.

We had a load of places that constructed Lsp.Error.t objects. They often had concatenate together stacks - the remote stack that was the root of the failure, and the local stack where we picked it up.

This diff uses a centralized function to construct Lsp.Error.t objects. It gets rid of a load of ad-hoc methods in Lsp_fmt that shouldn't have been there. It preserves stacks in structured json rather than using string concatenation.

Reviewed By: arxanas

Differential Revision: D21150066

fbshipit-source-id: 81903e56fdc11bfce3bf6101536ce90892eb320d