re-order methods
Summary:
This diff is nothing other than moving methods around. I created a section of the file to do with "connection to servers".

(why? because I ultimately want ide_rpc to be able to update status. But updating status involves closing over enough functions and state to be able to restart the servers, in case the user clicks the "restart" button. So, everything concerning starting the servers had to be moved above ide_rpc).

Reviewed By: arxanas

Differential Revision: D21115572

fbshipit-source-id: 64dec6d61c7fef0027eb69a1c5c9559fabc2d8ce