refactor clientIde errors
Summary:
This diff pays off some tech debt I'd accumulated.

Things were a bit muddled before. This diff makes it clear that we use `stopped_reason` to store human-facing data, and `Lsp.Error.t` to store programmatic/telemetry data. I picked Lsp.Error.t because it stores exactly the right things (a message and a structure json), and because anyway one of the two places we write telemetry is to jsonrpc errors. We sometimes need both - e.g. when load-state fails, then we need to report the human-facing error immediately, and we need to store up the telemetry so it can be written on every subsequent request (to correlate future failures with their root cause).
* We used to report telemetry event SERVERLESS_IDE_CRASH for one case of a bug that results in the ide service shutting down. I generalized that to SERVERLESS_IDE_BUG which records all manner of bugs, the bugs whose telemetry doesn't flow to clientLsp and hence doesn't appear in normal per-message clientLsp telemetry.
* Failed_init should take Lsp.Error.t as a parameter (not the human-facing stopped_reason) since it's only ever used for subsequent telemetry.
* I changed ClientIdeDaemon and ClientIdeService systematically to share some common routines for all cases where they report bugs. This way, I no longer need a zillion different places that fetch the exception stack.
* ClientIdeMessage.Stopped no longer takes an explicit Stop_reason.t. Instead, that gets subsumed into the richer-structured Lsp.Error.t for telemetry, that includes callstack as well as stop-reason.
* I changed ClientIdeMessage.stop. Formerly, if "stop" was called, that was logged in SERVERLESS_IDE_CRASH telemetry event, but didn't show up in per-message telemetry. Hence I had do to an irritating manual cross-telemetry-row correlation to find the root cause of a message failure. Now it carries the telemetry (callstack + stop_reason) in the stopped state. I also distinguished the stop_reason of "crashed" vs "stopped normally". I really want telemetry about "Stopped" state to be easier to investigate.

Reviewed By: arxanas

Differential Revision: D21148914

fbshipit-source-id: 06f450861fa5c23366cf4e5f719e252268454fc9