handle documentSymbol before init
Summary:
textDocument/documentSymbols is what powers Outline. It's typically sent as the very first request after initialize.

Hitherto we've always rejected the request because we were busy initializing, busy loading the saved-state. And so the user got a "no outline" message in the outline view. They only got an outline if they (1) closed the file, (2) waited for Hack to finished loading saved state, (3) re-opened the file.

There's no need for that. Outline is a purely syntactic thing that doesn't require saved-state. This diff lets us show it immediately.

Reviewed By: arxanas

Differential Revision: D21107617

fbshipit-source-id: f637b82e37bd789af28d141e146764b4a21f70a3