ide_service ref option
Summary:
Previously we used to have `ide_server : ClientIdeService.t option ref`. That indicated that it's a mutable reference which might get switched from None to Some or vice versa.

But that's not true. Really it should be `ide_server : ClientIdeService.t ref option`. In other words, either it starts as None and shall remain that forever, or it starts as Some and shall remain that forever (although it might get mutated each time the user restarts HackIDE).

That's what this diff does.

Note that ide_rpc needs to take a mutable `ClientIdeService.t ref`. That's because it will report status, and reporting status closes over the possibility of the user hitting restart, an in that case the ref will mutate. Indeed, mutability of `ClientIdeService.t ref` is pervasive.

Reviewed By: arxanas

Differential Revision: D21119732

fbshipit-source-id: 39d31c69fd074b82ba20551f3d251baaa5c0af5a