tighten up status updates
Summary:
The way we do status updates is complicated and mysterious.

I've added a lot of comments.

I also moved "latest_hh_server_status" out of a local variable and into a global variable. That's because I have the ultimate goal that we should be able to refresh status without needing the state monad, e.g. during an RPC.

Reviewed By: arxanas

Differential Revision: D21114281

fbshipit-source-id: 7e364449a4b9fbc84d2d41da5651a040c186a0dc