Use get instead of rvalVanilla in JIT getters
Summary:
As alexeyt pointed out, these JIT helpers are logically reading an array value, so we don't need to use the tv_rval const-ref type here.

Using get may cause us to generate another branch here, because get uses rval internally and branches on whether it's null, then we branch again on whether the TypedValue is init. As long as perf overall is flat, we're okay. In fact, this change looks like a small (~0.1%) load-store win.

Reviewed By: ricklavoie

Differential Revision: D21197185

fbshipit-source-id: ffcc97e67e336f50c71bdbc30f5e90db157edc0f