Make false-y promotion errors consistent
Summary:
Right now, if a class property has type ?int, then triggering false-y promote for it throws a type-hint violation error "Cannot assign array to int property", but if the class property has type ?darray, triggering false-y promote for it throws a false-y promote exception.

This diff modifies our logic so that we consistently emit the false-y promote exception. Doing so will allow us to eliminate MInstrPropState completely. I'll do the DCE in a follow-up diff, since this diff is a small one that changes behavior and that diff is a big one that does not change behavior.

Reviewed By: ricklavoie

Differential Revision: D21213917

fbshipit-source-id: 70a3c2607b77fc726c8816e3c9468dd1b761c389