3/12 don't use lazy in Naming_sqlite
Summary:
This diff stack gets rid of global sharedmem state "NamingTableDatabaseSettings", which stores the path to the naming DB.

In this diff, we had been using "lazy" to represent the handle to the Sqlite3.db. Lazy is a nice thing, sure, but it's getting in the way of my ability to refactor the code. So I'm refactoring it to no longer use Lazy.

Reviewed By: arxanas

Differential Revision: D21191770

fbshipit-source-id: c7a7a2b993e2ef8798fb438d80653af8546b626b