4/12 cache db_path
Summary:
This diff stack gets rid of global sharedmem state "NamingTableDatabaseSettings", which stores the path to the naming DB.

The existing code worked fine. It used that global sharedmem state for the db_path, and it cached the reference to the Sqlite3.db in ocaml worker memory.

But I want to make Naming_sqlite stateless and purely functional. So, I don't want to use the cached reference to Sqlite3.db semantic information; instead I'll have each function take in a db_path themselves.

Where will they fetch db_path? Ultimately it will be provided in an efficient way by the context. But I'll only get there incrementally one step at a time.

For now, I'll just add a local ocaml worker cache for the sharedmem "NamingTableDatabaseSettings". This will give us the freedom to retrieve db_path as often as we like with no perf hit.

(Last time I attempt this diffstack, it had a considerable perf hit. My current conjecture is that the perf hit came from repeatedly hitting SharedMem for the db_path, and uncompressing+unmarshalling it every single time. That's why I want to cache it. I might concievably be able to use SharedMem.WithCache, but that doesn't cache absences, so it would introduce a perf degradation. Maybe that perf degradation is okay? it'd need future study in a separate diff.)

Reviewed By: arxanas

Differential Revision: D21191987

fbshipit-source-id: fe75e99bcd785a8c11751c9194181dd6ad9dcde3