10/12 Naming_table.Backed stores db_path
Summary:
This diff stack gets rid of global sharedmem state "NamingTableDatabaseSettings", which stores the path to the naming DB.

This diff isn't great. It introduces redundant storage of the db_path. But it's the best I could manage, and I think is worth it to achieve the goal of eliminating SharedMem. Here's why.

1. We use Naming_table.t to store the forward naming table delta, and use Naming_table as the API to retrieve forward names (via the delta and then by the backing sqlite database).
2. We also using Naming_table as the API to set up the backing sqlite database.
3. We use Provider_backend to store the reverse  naming table delta, and use Naming_provider to retrieve reverse names (via the delta and then by the backing sqlite database)

The setup of the sqlite database is shared. The fact that it's shared is represented by a side-effect: Naming_table's internals write into global sharedmem "db_path", and subsequently Naming_provider's internals reads from it. Yuck.

I wanted the single source of truth for db_path to be stored inside Provider_context. But this turned out quite tricky.
* That's because Naming_table.t's initialization needs the path, and it is part of `env`, and we need an initialized `env` before we can construct a Provider_context.
* I tried to make it so Provider_context is constructed with an already-prepared Naming_table.t, i.e. fixing the db_path early up front. This was tricky for the same reasons.
* There are a heck of a lot of callsites which use Naming_table.t but which don't yet have Provider_context threaded through to them. And many of those callsites will never be ported to new architecture of decl-service or serverless-IDE. So I didn't want to invest time in changing them to take Provider_context.t.

In all, it felt slickest to make `Naming_table.Backed` store the db_path itself. After all, the definition of `Backed` is that it has a db_path! This means I don't need to change many callsites at all. I only need to change the single callsite that constructs the `Backed`.

Reviewed By: arxanas

Differential Revision: D21192653

fbshipit-source-id: 865067efc9683a8e6d27629d8d69dc97f362bfaf