12/12 push Naming_sqlite.db_path up to its callers
Summary:
This diff stack gets rid of global sharedmem state "NamingTableDatabaseSettings", which stores the path to the naming DB.

As discussed in the previous diff, db_path is stored in three places: (1) inside the Provider_backend, (2) that gets copied and stored inside Naming_table.Backed, (3) that gets copied and written via Naming_sqlite into sharedmem.

With this diff, we eliminate (3) from Naming_sqlite. All callers of Naming_sqlite already have at their disposal either (1) or (2). And so with this diff, that's the db_path they pass into Naming_sqlite.

Reviewed By: arxanas

Differential Revision: D21192957

fbshipit-source-id: a5f141b1e79873c7a4deaf3a0ede386a2377f2a5