2/12 refactor Naming_table.save
Summary:
This diff stack gets rid of global sharedmem state "NamingTableDatabaseSettings", which stores the path to the naming DB.

In this diff, Naming_table.save used to have too much visibility into that global state, and mutated it along the way. Needlessly. This diff achieves the same "save" but without mutating global state.

Reviewed By: arxanas

Differential Revision: D21178207

fbshipit-source-id: c1b790688200dfb078644d8a7a2600ad3289af43