11/12 no need to unset db_path
Summary:
This diff stack gets rid of global sharedmem state "NamingTableDatabaseSettings", which stores the path to the naming DB.

We store db_path in two places: (1) inside the Provider_backend, (2) that gets copied and stored inside Naming_table.Backed. (3) for now, it also gets copied and stored inside sharedmem, but the goal of my diffstack is to eliminate (3).

If we've got an unbacked naming-table in hand, and we save it, then prior to this diff the code used to unset (3). That's confusing. If we didn't have a Backed Naming_table.t, then we wouldn't have written any db_path into (3) in the first place, and indeed (1) presumably didn't have a db_path either.

I'm making this change because I wish to eliminate (3), and the method Naming_table.save doesn't have access to unset (1), and I don't think it has any need to, and I don't want to thread the ability to do so through all of its many callers.

Reviewed By: arxanas

Differential Revision: D21192850

fbshipit-source-id: 409ec9480c8d0ec5ea2d26d53f7c8fa90eb32b1c