7/12 thread db_path down through Naming_provider
Summary:
This diff stack gets rid of global sharedmem state "NamingTableDatabaseSettings", which stores the path to the naming DB.

The Provider_context already stores the path to the naming DB. That's where we want it (not in mutable sharedmem storage). We just need to thread it through from the high level that knows ctx, down to the lower levels that call into Naming_sqlite.

This diff pushes it down level, inside the functions of Naming_provider. They will use it by the end of this stack.

Reviewed By: arxanas

Differential Revision: D21192210

fbshipit-source-id: 6bf29b353366a9172297242edf732824e533e4de