coverage shouldn't swallow exceptions
Summary:
Coverage was swallowing exceptions. This diff fixes that.

People generally used to invoke it
```
Fixmes.HH_FIXMES.find_unsafe
```
which threw an exception if the fixme-map wasn't found for the file

D15169752 switched over to Fixme_provider and preserved the exact same behavior.

Here's what the coverage caller looked like:
```
  try
    let { Tast_provider.Compute_tast.tast; _ } =
      Tast_provider.compute_tast_quarantined ~ctx ~entry
    in
    Coverage_level.get_levels ctx tast entry.Provider_context.path
  with _ ->
    (* The "Fixme Provider" will throw an exception if the file cannot be found.
     * Let's convert that exception to a plain result. *)
```

This isn't nice! It means that exceptions in our fixme provider will be silently swallowed, both the expected ones (fixmes-map not found) and unexpected ones (failwith "not implemented")

This diff switches to using a result: Ok for the normal case, Error for the fixmes-map-not-found case, and we'll still have exceptions for unexpected behavior.

Reviewed By: arxanas

Differential Revision: D21207509

fbshipit-source-id: 682cdac644cad1e05c240f94090454f489675c17