Add `_OS\request_stdio_fd()`
Summary:
- allows retrieving the raw FD for CLI request stdin/stdout/stderr
- fails with EBADF on any other file handle
- fails with EBADF in non-CLI modes (e.g. proxygen, fastcgi)

`OS\` is unlikely to expose in/out for non-CLI modes, as there is not a safe
per-request fd; instead, I'll need to add a new API to expose request/response
on HPHP::Transport; that will be used by `IO\` directly, not via `OS\`

This is done by allowing `HSLFileDescriptor` to wrap PlainFile resources (which
are FD-based - and in-turn, have a FILE*, which have an FD).

- We need to wrap rather than dupe so that `\close(STDOUT)` and
  `_OS\close($stdout_fd)` are largely equivalent, and affect each other
- I am not fetching the FD once, then storing it in the HSLFileDescriptor, as
  the HSLFileDescriptor must be invalidated when the PHP stream is closed
- Ditto for the FILE*; additionally, if hhvm is invoked with `--count n`,
  PlainFile would attempt to reset the buffer for the same FILE*, which would
  fail
- so... we need to keep the resource, and operate on it every time.

Adding a union meant that default generated constructors and destructors no
longer work,however NativeData requires a 0-arg constructor, so added one.

Moved the existing logic to overload constructors, and made the factory method
use placement new to consistently use constructors for all initialization.

Reviewed By: alexeyt

Differential Revision: D21031958

fbshipit-source-id: 7e94c25fc9d9e6df792c1b6b84393b9b8de15e41