Remove PromoteEmptyObject flag
Summary:
This flag is always set to false except in a few tests. Remove it and apply DCE. The end result is that a bunch of final member ops (SetProp, SetOpProp, and IncDecProp) no longer need to take an MInstrPropState pointer.

billf made the same change in D19240695. By pushing through this change plus a few more on top of it, we can eliminate MInstrPropState completely.

Reviewed By: billf

Differential Revision: D21213520

fbshipit-source-id: 8fe2168a952f222931599fed1b6eb500775b681d