Merged future should not short-cirtuit on first future's is_ready result
Summary:
Calling `is_ready` has the side effect of advancing progress of bound futures.

Therefore, if we merge 2 bound futures, we should be checking both whether they are ready when `is_ready` on the merged future is called, to ensure that the 2 futures do as much work in parallel as possible.

Reviewed By: yinghuitan

Differential Revision: D21197354

fbshipit-source-id: 07c7a927d6943b133324cf740fd4f7c34fd8345c