Add flag to ban non constant default arguments on lambdas
Summary:
This diff adds a parser flag to make non constant initializers in default parameters for lambdas (and php style anonymous functions) a parse error.

The motivation for this comes from two parts:
- We are already making progress towards banning non constant initializers for default parameters and it is a goal that we are actively working towards.
- We have some niche bugs that have burned people regarding non constant initializers in lambdas particularly

We already have an existing flag that bans all non constant intializers in default parameters (not just in anonymous functions or lambdas) and this flag will throw a subset of that flag's errors. (When both flags are turned on, non const default parameters will throw the same error twice.)

This is such that it is possible to migrate the codebase in pieces - I imagine that it will be easier to remove this behavior from lambdas first. When we finally are able to flip on the full flag, we can remove the lambda only flag.

Reviewed By: shiqicao

Differential Revision: D21079073

fbshipit-source-id: 25066603603d202625067f0e623e25e278e4d720