Enable like casts to non-enforceable generics
Summary: Initially we held off on allowing arbitrary like casts to generics to match reified generics' restriction on casting to callables etc. However, one can always cast to dynamic and pipe a value through a function that returns the desired like type, so that's the same as just allowing the like cast in the first place.

Reviewed By: DavidSnider

Differential Revision: D21237630

fbshipit-source-id: c5f82898833681c23b5119b190d735222aa594df