Make ElemX a value-type getter
Summary: As with the previous diff, we'll make this method return a TypedValue rather than a ref type. ElemX is incredibly cold: I see a few usages in hundreds of thousands of translations. This diff eliminates the penultimate "as_lval" usage; I'll pick off the last one in the next diff, which does DCE now that so many Elem IR ops are dead.

Reviewed By: ricklavoie

Differential Revision: D21217242

fbshipit-source-id: 6d2aba234d5f2765080944448f1f65f27406296a