Implement post-specialization comparison behavior
Summary:
The post-specialization behavior is the same as what we currently do for vecs and dicts, modulo error messages:
- For Eq and Same with mismatched dvarray bits, we'll return false.
- For relational comparisons between two darrays, we'll throw.
- For relational comparisons with mismatched dvarray bits, we'll throw an error based on one of ad1's or ad2's types.

There is a bit of subtlety hidden in this diff. Specifically, by un-setting the notice flag when the specialization is on, we change HHBBC's analysis to mark more bytecode ops as "no throw". We need to check that none of the ops that still throw after specialization is on are marked "no throw". I've done this audit; the result is that we'll mark Same/NSame/Eq/NEq as no throw, but we still don't mark the comparisons.

Reviewed By: ricklavoie

Differential Revision: D21237303

fbshipit-source-id: be51e428f51a987107a51d80ddbdb269cfabb130